requirements.txt
PyQt6>=6.5
cryptography>=41.0
argon2-cffi>=21.3
pandas>=2.0.1  # optional, useful for export features
